app/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sse-starlette==1.8.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis and Celery
redis==5.0.1
celery==5.3.4

# MinIO / S3
minio==7.2.0
boto3==1.29.7

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Video Processing
ffmpeg-python==0.2.0
openai-whisper==20231117

# Monitoring
prometheus-client==0.19.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
